Complete Project For Ever !!!!!!!!!!!!!!!!!!!!

+++ app/Http/Controllers/BooksController.php
@@ -126,12 +126,24 @@ class BooksController extends Controller
      */
     public function destroy(string $id)
     {
-        try {
-            Book::find($id)->delete();
-            return response()->json(['status' => 200, 'message' => 'کتاب باموفقیت حذف شد .']);
-        } catch (QueryException $exception) {
-            return response()->json(['status' => 201, 'message' => 'خطا در حذف کتاب ، مجددا تلاش کنید .']);
-        }
+        if ($this->checkBookReserved($id)) {
+            try {
+                Book::find($id)->delete();
+                return response()->json(['status' => 200, 'message' => 'کتاب باموفقیت حذف شد .']);
+            } catch (QueryException $exception) {
+                return response()->json(['status' => 201, 'message' => 'خطا در حذف کتاب ، مجددا تلاش کنید .']);
+            }
+        } else
+            return response()->json(['status' => 202, 'message' => 'کتاب در دست امانت است !']);
+    }
+
+    private function checkBookReserved($id)
+    {
+        $book = Book::find($id);
+        if ($book)
+            if ($book->status == 0)
+                return true;
+        return false;
     }
 
     public function search(Request $request)
@@ -139,6 +151,9 @@ class BooksController extends Controller
         $word = $request->word;
         $books = Book::orWhere('title', 'like', '%' . $word . '%')
             ->orWhere('barcode', 'like', '%' . $word . '%')->paginate(20);
+
+        if ($request->has('isBookPage'))
+            return view('backend.books.partial', compact('books'));
         return view('backend.users.books_partial', compact('books'));
     }
 
@@ -149,7 +164,7 @@ class BooksController extends Controller
             $data = [
                 'user_id' => Auth::id(),
                 'book_id' => $id,
-                'reserve_date' => Carbon::now() ,
+                'reserve_date' => Carbon::now(),
                 'period' => 7
             ];
             if ($this->createReserve($data))
